feat(middlewares): change middlewares to become stateless
The middlewares were providing a statefull logic. The issue was that each middleware was applied for
the pubsub. So no matter how many stores there was the middlewares were one for all of them. Now the
middlewares are based on the provided models

BREAKING CHANGE: The middlewares api is totally changed